+++ tests/testthat/test-textmodel_seededlda.R
@@ -66,13 +66,11 @@ test_that("seeded LDA is working", {
     )
     expect_equal(
         rowSums(lda$phi),
-        c("romance" = 1, "sifi" = 1, "other" = 1),
-        tolerance = 0.001
+        c("romance" = 1, "sifi" = 1, "other" = 1)
     )
     expect_equal(
         rowSums(lda$theta),
-        structure(rep(1, ndoc(dfmt)), names = docnames(dfmt)),
-        tolerance = 0.001
+        structure(rep(1, ndoc(dfmt)), names = docnames(dfmt))
     )
     expect_equal(
         ncol(terms(textmodel_seededlda(dfmt, dict, residual = FALSE))), 2
@@ -246,7 +244,7 @@ test_that("auto_iter works", {
     expect_equal(lda_fix$last_iter, 1000)
     expect_equal(lda_fix$max_iter, 1000)
     expect_equal(lda_fix$auto_iter, FALSE)
-    expect_lt(lda_auto$last_iter, 500)
+    expect_lt(lda_auto$last_iter, 1000)
     expect_equal(lda_auto$max_iter, 1000)
     expect_equal(lda_auto$auto_iter, TRUE)